+++ src/settings.js
@@ -157,7 +157,8 @@ var MALLOC = "dlmalloc";
 // previously. If you don't want that, just stop passing it in at link time.
 //
 // Note that this setting does not affect the behavior of operator new in C++.
-// This function will always abort on allocation failure if exceptions are disabled.
+// This function will always abort on allocation failure if exceptions are
+// disabled.
 // If you want new to return 0 on failure, use it with std::nothrow.
 //
 // [link]
@@ -276,8 +277,8 @@ var INITIAL_TABLE = -1;
 var ALLOW_TABLE_GROWTH = false;
 
 // Where global data begins; the start of static memory.
-// A GLOBAL_BASE of 1024 or above is useful for optimizing load/store offsets, as it
-// enables the --low-memory-unused pass
+// A GLOBAL_BASE of 1024 or above is useful for optimizing load/store offsets,
+// as it enables the --low-memory-unused pass
 // [link]
 var GLOBAL_BASE = 1024;
 
@@ -402,16 +403,16 @@ var FS_DEBUG = false;
 // [link]
 var SOCKET_WEBRTC = false;
 
-// A string containing either a WebSocket URL prefix (ws:// or wss://) or a complete
-// RFC 6455 URL - "ws[s]:" "//" host [ ":" port ] path [ "?" query ].
-// In the (default) case of only a prefix being specified the URL will be constructed from
-// prefix + addr + ':' + port
+// A string containing either a WebSocket URL prefix (ws:// or wss://) or a'
+// complete RFC 6455 URL - "ws[s]:" "//" host [ ":" port ] path [ "?" query ].
+// In the (default) case of only a prefix being specified the URL will be
+// constructed from prefix + addr + ':' + port
 // where addr and port are derived from the socket connect/bind/accept calls.
 // [link]
 var WEBSOCKET_URL = 'ws://';
 
-// If 1, the POSIX sockets API uses a native bridge process server to proxy sockets calls
-// from browser to native world.
+// If 1, the POSIX sockets API uses a native bridge process server to proxy
+// sockets calls from browser to native world.
 // [link]
 var PROXY_POSIX_SOCKETS = false;
 
@@ -484,12 +485,13 @@ var GL_EXTENSIONS_IN_PREFIXED_FORMAT = true;
 // [link]
 var GL_SUPPORT_AUTOMATIC_ENABLE_EXTENSIONS = true;
 
-// If true, the function ``emscripten_webgl_enable_extension()`` can be called to
-// enable any WebGL extension. If false, to save code size,
-// ``emscripten_webgl_enable_extension()`` cannot be called to enable any of extensions
-// 'ANGLE_instanced_arrays', 'OES_vertex_array_object', 'WEBGL_draw_buffers',
-// 'WEBGL_multi_draw', 'WEBGL_draw_instanced_base_vertex_base_instance',
-// or 'WEBGL_multi_draw_instanced_base_vertex_base_instance',
+// If true, the function ``emscripten_webgl_enable_extension()`` can be called
+// to enable any WebGL extension. If false, to save code size,
+// ``emscripten_webgl_enable_extension()`` cannot be called to enable any of
+// extensions 'ANGLE_instanced_arrays', 'OES_vertex_array_object',
+// 'WEBGL_draw_buffers', 'WEBGL_multi_draw',
+// 'WEBGL_draw_instanced_base_vertex_base_instance', or
+// 'WEBGL_multi_draw_instanced_base_vertex_base_instance',
 // but the dedicated functions ``emscripten_webgl_enable_*()``
 // found in html5.h are used to enable each of those extensions.
 // This way code size is increased only for the extensions that are actually used.
@@ -498,9 +500,9 @@ var GL_SUPPORT_AUTOMATIC_ENABLE_EXTENSIONS = true;
 // [link]
 var GL_SUPPORT_SIMPLE_ENABLE_EXTENSIONS = true;
 
-// If set to 0, Emscripten GLES2->WebGL translation layer does not track the kind
-// of GL errors that exist in GLES2 but do not exist in WebGL. Settings this to 0
-// saves code size. (Good to keep at 1 for development)
+// If set to 0, Emscripten GLES2->WebGL translation layer does not track the
+// kind of GL errors that exist in GLES2 but do not exist in WebGL. Setting
+// this to 0 saves code size. (Good to keep at 1 for development)
 // [link]
 var GL_TRACK_ERRORS = true;
 
@@ -1483,8 +1485,8 @@ var EMIT_EMSCRIPTEN_LICENSE = false;
 
 // Whether to legalize the JS FFI interfaces (imports/exports) by wrapping them
 // to automatically demote i64 to i32 and promote f32 to f64. This is necessary
-// in order to interface with JavaScript.  For non-web/non-JS embeddings, setting
-// this to 0 may be desirable.
+// in order to interface with JavaScript.  For non-web/non-JS embeddings,
+// setting this to 0 may be desirable.
 // [link]
 // [deprecated]
 var LEGALIZE_JS_FFI = true;
@@ -2107,9 +2109,9 @@ var ERROR_ON_WASM_CHANGES_AFTER_LINK = false;
 var ABORT_ON_WASM_EXCEPTIONS = false;
 
 // Build binaries that use as many WASI APIs as possible, and include additional
-// JS support libraries for those APIs.  This allows emscripten to produce binaries
-// are more WASI compliant and also allows it to process and execute WASI
-// binaries built with other SDKs (e.g.  wasi-sdk).
+// JS support libraries for those APIs.  This allows emscripten to produce
+// binaries that are more WASI compliant and also allows it to process and
+// execute WASI binaries built with other SDKs (e.g.  wasi-sdk).
 // This setting is experimental and subject to change or removal.
 // Implies :ref:`STANDALONE_WASM`.
 // [link]
@@ -2135,7 +2137,7 @@ var IMPORTED_MEMORY = false;
 // with the ``.orig`` suffix and one without.  The default file (without
 // the suffix) when run will generate instrumentation data can later be
 // fed into wasm-split (the binaryen tool).
-// As well as this the generated JS code will contains help functions
+// As well as this the generated JS code will contains helper functions
 // to loading split modules.
 // [link]
 // [experimental]
@@ -2150,9 +2152,10 @@ var AUTOLOAD_DYLIBS = true;
 // though these syscalls will fail (or do nothing) at runtime.
 var ALLOW_UNIMPLEMENTED_SYSCALLS = true;
 
-// Allow calls to Worker(...) and importScripts(...) to be Trusted Types compatible.
-// Trusted Types is a Web Platform feature designed to mitigate DOM XSS by restricting
-// the usage of DOM sink APIs. See https://w3c.github.io/webappsec-trusted-types/.
+// Allow calls to Worker(...) and importScripts(...) to be Trusted Types
+// compatible. Trusted Types is a Web Platform feature designed to mitigate DOM
+// XSS by restricting the usage of DOM sink APIs.
+// See https://w3c.github.io/webappsec-trusted-types/.
 // [link]
 var TRUSTED_TYPES = false;
 
@@ -2186,21 +2189,22 @@ var RUNTIME_DEBUG = 0;
 
 // Include JS library symbols that were previously part of the default runtime.
 // Without this, such symbols can be made available by adding them to
-// :ref:`DEFAULT_LIBRARY_FUNCS_TO_INCLUDE`, or via the dependencies of another JS
-// library symbol.
+// :ref:`DEFAULT_LIBRARY_FUNCS_TO_INCLUDE`, or via the dependencies of another
+// JS library symbol.
 var LEGACY_RUNTIME = false;
 
-// User-defined functions to wrap with signature conversion, which take or return
-// pointer argument. Only affects ``MEMORY64=1`` builds, see
+// User-defined functions to wrap with signature conversion, which take or
+// return pointer arguments. Only affects ``MEMORY64=1`` builds, see
 // ``create_pointer_conversion_wrappers`` in ``emscripten.py`` for details.
-// Use _ for non-pointer arguments, p for pointer/i53 arguments, and P for optional pointer/i53 values.
+// Use ``_`` for non-pointer arguments, ``p`` for pointer/i53 arguments, and
+// ``P`` for optional pointer/i53 values.
 // Example use ``-sSIGNATURE_CONVERSIONS=someFunction:_p,anotherFunction:p``
 // [link]
 var SIGNATURE_CONVERSIONS = [];
 
 // Experimental support for wasm source phase imports.
-// This is only currently implemented in the pre-release/nightly version of node,
-// and not yet supported by browsers.
+// This is only currently implemented in the pre-release/nightly version of
+// node, and not yet supported by browsers.
 // Requires EXPORT_ES6
 // [link]
 // [experimental]