+++ tools/building.py
@@ -282,10 +282,7 @@ def lld_flags_for_executable(external_symbols):
   return cmd
 
 
-def link_lld(args, target, external_symbols=None):
-  if not os.path.exists(WASM_LD):
-    exit_with_error('linker binary not found in LLVM directory: %s', WASM_LD)
-  # runs lld to link things.
+def lld_flags(args):
   # lld doesn't currently support --start-group/--end-group since the
   # semantics are more like the windows linker where there is no need for
   # grouping.
@@ -305,29 +302,34 @@ def link_lld(args, target, external_symbols=None):
     # is passed.
     args.append('--keep-section=target_features')
 
-  cmd = [WASM_LD, '-o', target]
+  if settings.MEMORY64:
+    args.append('-mwasm64')
+
   for a in llvm_backend_args():
-    cmd += ['-mllvm', a]
+    args += ['-mllvm', a]
 
   if settings.WASM_EXCEPTIONS:
-    cmd += ['-mllvm', '-wasm-enable-eh']
+    args += ['-mllvm', '-wasm-enable-eh']
     if settings.WASM_LEGACY_EXCEPTIONS:
-      cmd += ['-mllvm', '-wasm-use-legacy-eh']
+      args += ['-mllvm', '-wasm-use-legacy-eh']
     else:
-      cmd += ['-mllvm', '-wasm-use-legacy-eh=0']
+      args += ['-mllvm', '-wasm-use-legacy-eh=0']
   if settings.WASM_EXCEPTIONS or settings.SUPPORT_LONGJMP == 'wasm':
-    cmd += ['-mllvm', '-exception-model=wasm']
+    args += ['-mllvm', '-exception-model=wasm']
 
-  if settings.MEMORY64:
-    cmd.append('-mwasm64')
+  return args
 
+
+def link_lld(args, target, external_symbols=None):
+  # runs lld to link things.
+  if not os.path.exists(WASM_LD):
+    exit_with_error('linker binary not found in LLVM directory: %s', WASM_LD)
+  cmd = [WASM_LD, '-o', target]
   # For relocatable output (generating an object file) we don't pass any of the
   # normal linker flags that are used when building and executable
   if '--relocatable' not in args and '-r' not in args:
     cmd += lld_flags_for_executable(external_symbols)
-
-  cmd += args
-
+  cmd += lld_flags(args)
   cmd = get_command_with_possible_response_file(cmd)
   check_call(cmd)